chore: refactor iOS components to add extra
