+++ example/src/Views/Checkout/CheckoutView.js
@@ -23,9 +23,10 @@ const CheckoutView = ({ navigation }) => {
   const didSubmit = useCallback(
     async (
       /** @type {import('@adyen/react-native').PaymentMethodData} */ data,
-      /** @type {import('@adyen/react-native').AdyenActionComponent} */ nativeComponent
+      /** @type {import('@adyen/react-native').AdyenActionComponent} */ nativeComponent,
+      /** @type any */ extra
     ) => {
-      console.log(`didSubmit: ${data.paymentMethod.type}`);
+      console.log(`didSubmit: ${data.paymentMethod.type} with extra: ${JSON.stringify(extra, null, " ")}`);
       try {
         /** @type {import('@adyen/react-native').PaymentResponse} */
         const result = await ApiClient.payments(data, configuration);
@@ -130,6 +131,24 @@ const CheckoutView = ({ navigation }) => {
               applepay: {
                 merchantID: ENVIRONMENT.applepayMerchantID,
                 merchantName: configuration.merchantName,
+                billingContact: {
+                  phoneNumber: 'phoneNumber',
+                  emailAddress: 'emailAddress',
+                  givenName: 'name',
+                  familyName: 'familyName',
+                  addressLines: ['addressLines1'],
+                  locality: 'locality',
+                  subLocality: 'subLocality',
+                  subAdministrativeArea: 'subAdministrativeArea',
+                  postalCode: 'postalCode',
+                  administrativeArea: 'administrativeArea',
+                  country: 'country',
+                  countryCode: 'countryCode',
+                  phoneticGivenName: 'phoneticName',
+                  phoneticFamilyName: 'phoneticFamilyName'
+                },
+                requiredBillingContactFields: [ 'name', 'postalAddress' ],
+                requiredShippingContactFields: [ 'name', 'postalAddress' ]
               },
             }
           }
