+++ ios/Components/BaseModule.swift
@@ -22,7 +22,7 @@ internal class BaseModule: RCTEventEmitter {
     override static func requiresMainQueueSetup() -> Bool { true }
     override func stopObserving() { /* No JS events expected */ }
     override func startObserving() { /* No JS events expected */ }
-    override open func supportedEvents() -> [String]! { Events.allCases.map(\.rawValue) }
+    override open func supportedEvents() -> [String]! { [] }
 
     internal var currentComponent: Component?
     internal var currentPaymentComponent: PaymentComponent? {
