+++ ios/Model/Error.swift
